[VEG-180] feat: delete redundant parts
